Add User Story Status Field

+++ src/main/java/edu/umich/kanboard/userstory/UserStoryRepository.java
@@ -8,7 +8,7 @@ public interface UserStoryRepository extends CrudRepository<UserStoryEntity, Lon
 
     List<UserStoryEntity> findAll();
 
-    UserStoryEntity findByUserStoryId(Integer userStoryId);
+    UserStoryEntity findByUserStoryId(Long userStoryId);
 
     List<UserStoryEntity> findTop8ByNameIgnoreCaseContainingOrderByName(String name);
